Bug #1187040: Random Segmentation faults on fileio test
Fixed a race condition in the test termination code with respect to
--report-interval: the reporting thread might read a non-zero
report_interval value and call print_stats(). If report_interval is
changed to 0 by the master thread and the test's done() method called,
print_stats() could access some shared resources deallocated in
done().

Fixed by protecting report_interval with a separate mutex which is
locked on updates and during the check by reporting thread and
print_stats().