Replace percentile mutex with CK atomics + rwlock.
The idea is to remove the mutex from the common path, i.e. worker
threads updating the percentile statistics. The only reason why an
rwlock is needed is sb_percentile_reset() in which case we want to block
all updates until the reset is done to not lose any events during the
reset process.

TODO: This can be further improved with ck_brlock and scalable counters.