Add caching to the SERIAL column.
Add timing information to a bunch of log lines to better understand the
time required for the various steps in loading.

Added a function enable_debug that sets the debug mode in yugabyte.

Test Plan:
Tested it manually against the cluster deployed locally and on portal.

Output for when we manually set the serial_cache_size:
   sysbench oltp_write_only --tables=1 --table-size=1000
  --db-driver=pgsql --pgsql-host=127.0.0.1 --pgsql-port=5433
  --pgsql-user=yugabyte --pgsql-db=yugabyte --threads=1
  --serial_cache_size=500 prepare
  sysbench 1.1.0-bd4b418 (using bundled LuaJIT 2.1.0-beta3)

  (14:23:19) Creating table 'sbtest1'...
  alter sequence with cache size: 500
  (14:23:20) Inserting 1000 records into 'sbtest1'
  (14:23:21) Creating a secondary index on 'sbtest1'...
  (14:23:21) Done Loading

Reviewers:
Neha, Karhik